Cultural Activity
Time Limit: None
Memory Limit: None
Dataset: Download the data for the problem from this link.
In this problem, we aim to analyze a dataset containing information about the most famous movies in the world. The dataset includes various details about movies, such as budget, revenue, cast, production companies, and genres. To enhance our cinematic knowledge, we have a few questions about this data, and you need to help find the answers from the dataset. For submission, write the answer to each question in a single line of a text file, similar to the example. Ultimately, upload a four-line text file named output.txt. If you don't know the answer to a question, write 0 in the corresponding line.
1) Consider the three genres with the highest number of movies. How many unique movies in total belong to these three genres? (Note: Some movies may belong to multiple genres.)
Write the answer to this question in the first line of the file as a number in English.
2) Understanding film production companies and their revenues is crucial in the film industry. Sort the production companies in descending order based on the number of movies they have produced. What is the total revenue earned by the fifth company in this list from its movies? (If a movie has multiple production companies, the revenue is divided equally among them.) Consider only the integer part of the answer and discard the decimal part.
Example: 737490204
Plain text
Write the answer to this question in the second line of the file as a number in English.
3) Sort the list of actors in descending order based on the number of movies they have acted in. If two actors have the same number of movies, sort them alphabetically. For the first actor in this list, which production company have they collaborated with the most, and in how many movies? Write your answer in the format shown below in the third line of the file.
Example: 'company name', 6
Plain text
4) Actors who have co-starred in more than six movies (at least seven) are considered artistic pairs. Find the artistic pairs in this list and calculate how many movies each pair has in common. What is the total number of shared movies across all artistic pairs? For example, if pair X, Y has 10 shared movies, pair E, X has 7 shared movies, and pair R, T has 8 shared movies, the final answer would be 7 + 8 + 10 = 25.
Write the answer to this question in the fourth line of the file as a number in English.
Each question is worth 35 points. Do not forget to write 0 in the corresponding line if you don't know the answer to a question. Answers are evaluated in order.